Initial DB implementation
Added module Postgresql.
Refactored some code.
Changed the way the bot gets channels and messages.
Bot also saves the logs message ID if it does not exists in the database. This might be changed to be created in the "on ready" section of the bot, so you dont have to create the messages manually.